ECO_MU/requirements.txt
absl-py==0.13.0
addict==2.2.1
albumentations==0.4.5
astor==0.8.0
astunparse==1.6.3
attributee==0.1.7
attrs==19.3.0
bidict==0.21.2
cached-property==1.5.2
cachetools==4.2.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Cython==0.29.14
data==0.4
decorator==4.4.2
dominate==2.6.0
easydict==1.9
flatbuffers==1.12
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.8
imagecorruptions==1.0.0
imageio==2.8.0
imageio-ffmpeg==0.4.5
imgaug==0.2.6
importlib-metadata==1.6.0
intel-openmp==2021.3.0
joblib==0.14.1
jpeg4py==0.1.4
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
latex==0.7.0
llvmlite==0.32.0
lmdb==1.2.1
loguru==0.6.0
lvis==0.5.3
Markdown==3.1.1
matplotlib==3.1.3
mkl==2021.3.0
mmcv==0.2.14
moviepy==1.0.3
networkx==2.4
numba==0.49.0
numpy==1.21.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
ordered-set==4.0.1
packaging==21.0
pandas==1.0.0
phx-class-registry==3.0.5
Pillow==7.0.0
pip==22.0.3
proglog==0.1.9
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.19
pydub==0.25.1
PyLaTeX==1.3.1
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.2.1
setuptools==58.0.4
Shapely==1.7.0
shutilwhich==1.1.0
six==1.15.0
sklearn==0.0
tb-nightly==2.8.0a20220109
tbb==2021.3.0
tempdir==0.7.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminaltables==3.1.0
torch==1.4.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.61.2
typing-extensions==3.7.4.3
urllib3==1.25.9
visdom==0.1.8.9
vot-toolkit==0.5.3
vot-trax==3.0.3
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
wget==3.2
wheel==0.36.2
wrapt==1.12.1
xlwt==1.3.0
yacs==0.1.8
zipp==3.1.0
